+++ src/components/agent/validators.ts
@@ -48,3 +48,28 @@ export const validateBankAccountUpdate = () => {
     alias: optionalString,
   });
 };
+
+export const validateDepositIndex = () =>
+  checkSchema({
+    id: {
+      in: ["params"],
+      isString: true,
+      optional: true,
+    },
+  });
+
+export const validateDepositUpdate = () =>
+  checkSchema({
+    id: {
+      in: ["params"],
+      isString: true,
+      isEmpty: false,
+      optional: false,
+    },
+    tracking_number: {
+      in: ["body"],
+      isString: true,
+      isEmpty: false,
+      optional: false,
+    },
+  });